Added lib/doscalls.c : This file collects all the calls that use dos_to_unix() to map filenames before accessing the UNIX filesystem.
The other changes are to make the code that previously allways called
the ambiguous functions (such as file_size(), that internally called
dos_to_unix()) to be unambiguous.

For example: file_size() becomes "dos_file_size()", that calls
dos_to_unix(), and file_size(), that does not.

It is now very explicit when we are calling a dos mapping call
and when we are not.

Also added string_to_sid unsigned fix by adding the strtoul from libg++
code into lib/replace.c and testing for strtoul in configure.in.

Jeremy.