HPUX fix. This looks like a big change but isn't actually, most of the changes are bzero -> memset as the compiler on HPUX is very picky about bzero (probably a strange define).
The significant changes are in configure.in, to fix the HPUX shadow.h
problem and also move the large file support to the correct place.

passdb/pass_check.c: Added includes for KRB5, DCE and AFS if requested in autoconf.

tests/trapdoor.c: Fix to use setresuid and setresgid if defined.

The rest of the changes are bzero -> memset.

Jeremy.