Updates to make SSL stuff work.
Modified configure.in to ensure that SSL related stuff gets into
configure.  Still a bit tacky, but a beginning.  We check to
ensure that if -with-ssl is called, that the SSL includes and libraries
are on the system or we fall over.

It seems to build OK under Linux.  Pity I could not have got it into
2.0.1