Added caching of user password if it was entered via getpass.  We were already caching the password if it came from the command line or from the environment. This completes the set and deals with the bloody nusance when reconnecting a connection after we have gone daemon...  Grrr...
Added code to clean up the mount point following a catastrophic failure
during reconnect.  The smbmount daemon was exiting but leaving the mount
point in an unusable state.  If smbmount must exit following a reconnect
failure, we "unmount" the mount point and clean up mnttab.  Currently,
the unmount works, fixing some really ugly I/O errors and failure when
trying to remount.  The cleanup on mnttab still has problems.  This
is better than what it was and doesn't break anything that wasn't broken
before, so I'm committing this in even with the mnttab problem.  Will
commit the fix to that when I figure out what is busted there...

	This brings the 2.0 branch smbmount program in line with the
main branch...

        -mhw-