3 *wonderful* bug fixes from Bob Boehmer <boehmer@worldnet.att.net>
In his own words :

The bugs are:

1. If a file "X" in the tarfile is an exact multiple of 512 bytes in length
then the extraction will fail for the file following "X" in the tarfile.
This is because get_file exits with the tar buffer pointing at the first
block of the next file if the current file was an exact multiple of 512
bytes (the tar buffer should be pointing at the final block of the file just
extracted).

2. If the tarfile is being fed through a pipe to smbclient then next_block
does not always work because fewer than the requested number of bytes may be
returned by the read call in next_block even though the pipe is still open.

3. If a file "X" in the tarfile is 0 bytes in length then the extraction
will fail for the file following "X" in the tarfile because next_block is
called prior to the call to get_file.

This might explain the smbtar problems people have been having
in the 2.x code.

Jeremy.