removed the following dce/rpc code from samba 2.0.x branch:
- smbd (\PIPE\samr)
- smbd (\PIPE\winreg)
- smbd (\PIPE\lsarpc - LsaLookupNames; LsaLookupSids)
- smbd (\PIPE\srvsvc - NetrFileEnum; NetrSessionEnum; NetrConnectionEnum)
- rpcclient
- rpctorture

where parsing and header files were common between rpcclient and smbd
these were also removed.  this is one-level only: functions that _these_
functions call, e.g smb_io_unistr3 in rpc_parse/parse_misc.c have _not_
been removed.

lib/util.c had two registry assistance functions removed because they
used #defines in include/rpc_reg.h.  this indicates that in main cvs
branch the #defines should actually be in smb.h.

the functionality removed is only that code which is necessary to get
USRMGR.EXE and SRVMGR.EXE to work.

functionality that is unaffected:

- browsing (net share enum)
- nt domain netlogons

functionality that will return to pre-nt-domain code (i.e. SMBtrans2):

- password changing


i am sorely tempted to remove the NETLOGON functionality however jeremy
is guaranteed to go absolutely mental if i do that (pre-existing users,
despite us not telling them 2.0 was ok to use as a PDC) and andrew is
probably going to remove my cvs access for this one as it is.

a good justification for removing NETLOGON functionality from 2.0 is so
that microsoft doesn't put up a KB article advising people not to use
samba because it damages NT clients.  with respect to 2.0.2, i agree
with them.  the fact that this is actually their fault is neither here
nor there: we shouldn't aggravate bugs in nt clients.