After a long chat with Luke (a pleasant one actually, thanks for asking :-), he is happy with me reverting the RPC change.
The reason I need this code in the 2.0 branch (even though it doesn't
work correctly) is I am using it as a stable area that no-one cares
about (as it doesn't work) to write the fully error checked version
of the RPC transport and parsing code.

Having it removed caused me a few problems :-).

Jeremy.