A small raft of fixes... :-).
Added the ability to announce as NT server or NT workstation to the
"announce as" parameter. Updated docs for change.

Added "min passwd length" parameter (and documented it) to allow minimum
unix password length on password change to be set at runtime. Default
is 5 characters set the local.h

Fixed the RPC code that was hardcoding the version at 5.2 and the
announce bitmask parameters.

Jeremy.