I finally wrote that smb filter program that Jeremy and I have talked about so often. It is a simple plug proxy that listens on port 139 and redirects connections to another machine on the same port. The nice thing is that you can filter the packets on the way through.
So, for example, if you want to see what NT server does with a NT
client when unicode isn't negotiated just set CAPABILITY_MASK and
CLI_CAPABILITY_MASK to CAP_UNICODE and run "smbfilter desthost
DESTNAME". Then connect to desthost with the NT client. The filter
will create a connection to desthost, set the destination netbios name
to DESTNAME in the session request and change the negotiated
capabilities on the connection to exclude unicode. Then you just fire
up a sniffer and you have a perfect record of how a NT server responds
to a NT client when it doesn't negotiate unicode.

There are lots of nice things we can do with this. I'm especially
looking forward to confirming some suspected bugs in NT clients by
reproducing them with a MS server with different negotiated
options.

to build it do "make smbfilter". The only documentation is the source
code and this commit message.