added a very simple netbench simulator to smbtorture. The tests are called NBW95 and NBWNT.
I don't have the right sort of hardware (100MBit etc) to test whether
the results of this similator are anything like what netbench sees,
but I'm hoping Jeremy will run that test for me.

A typical run would be:

smbtorture //server/share -Uuser%pass -o 10 -N 20 NBWNT NBW95

that would run a NT and a 95 client test with 20 clients and 10 files
per client. The result is given as seconds for the run.

I think that for best results you would use a few copies of smbtorture
running on different clients and each similating maybe 10 clients to
build up a large client load with a small number of client machines.