I was playing with smbmount and found that it didn't compile in the current 2.0 tree. Rather than patching it again I rewrote it to use the clientgen.c code. This means we can finally delete the crufty old clientutil.c code. Please honour a minutes silence for this code, it served us well and I hope it rests in peace ... outside our code tree.
the new smbmount works, and is *much* simpler than the old one. It
still needs some work as I used a rather heavy axe in writing it but
at least it is a better basis than what we had.

oh, yes. I also changed the smbmount syntax. The old syntax was too
ridiculous for words and only happened because of the way smbmount
came out of the interactive smbclient. Now the syntax is:

 smbmount //server/share mountpoint [options ...]

which is what it should be in order to match the normal mount
command. I'm sorry if this breaks scripts etc, but keeping the old
syntax would have been way too ugly. We can do a wrapper script if
necessary.