Support for arbitrary VFS options in smb.conf.  Options are of the format 'vfs option = name = value'.  It would be nice to get rid of the first equals sign, but that would involve modifications to the smb.conf parser.
There are a couple of naughties here - I've added a P_PTR enum as the
VFS options are parsed in param/loadparm.c and stored as a linked list
of structures.  I couldn't for the life of me write a FN_LOCAL_PTR()
macro to generate a lp_vfsoptions() function that actually worked so
for the moment it is coded by hand.
(This used to be commit 89ad85f0a3dfa233f6370ac6e3c97a1a645cfdfe)