improved code that shows how . and .. match as well as filenames, it for example:
Domain=[VNET] OS=[Windows NT 3.51] Server=[NT LAN Manager 3.51]
Domain=[VNET] OS=[Windows NT 4.0] Server=[NT LAN Manager 4.0]
+-- ++- mask=[\masktest\"] file=[\masktest\klik]
+-- ++- mask=[\masktest\>] file=[\masktest\ekbhmkcfgkc.agmkemja]
+-- ++- mask=[\masktest\>] file=[\masktest\agemlb]
+-- ++- mask=[\masktest\?] file=[\masktest\bhlfhmdekc]

shows that NT4.0 matches a .. directory with any single wildcard
character whereas NT3.51 doesn't. That is really weird.

the first + is a match on .
the 2nd + is a match on ..
the third + is a match on the filename