new uid handling code. After looking at the problem with AIX 4.1 I decided that the way we handled uid/euid is not good. I've now rewritten it to test properly for the 4 supported uid handling varients and I've added assert calls to ensure that the uid/euid are always set to hat we expect. This might possibly break some platforms (although I hope not) but I think that is better than the previous situation where we could silently do the wrong thing.
the disadvantage is that the full tsts are only possible when you run
configure as root. Otherise configure needs to make an educated guess.

the core of the new code is in lib/util_sec.c. Look there if you want
to know what is going on with this change.