OK.  This code works on a RedHat 6.0 system.  However smbpasswd time out of sending the session setup on Solaris 2.6.  No idea. I'll work on it some tomorrow.  This is to fix the "Unable to setup password vectors" thingy.
Also changed an inet_aton() to inet_addr() as the former is
not very portable :-)

Luke,  I set the redir flag to false because the connection to
the smb-agent was failing and smbpasswd bombed.  Double check me
on this one.



-jc
(This used to be commit e1d2b174caf5f0c48a8fac25778f72a868ec6eb7)