wow.  apart from what looks like an (unrelated) off-by-one bug in the calculation of a PDU, removing dce/rpc fragment reassembly from smbd actually worked!
all that's happening is an SMBread does a read on the socket; SMBwrite
does a write on the socket; SMBtrans does a write _and_ a read.

the rest is handled by the msrpc daemon, on the other side of the socket.
which of course, has this stupid off-by-one bug: i _just_ happened to
get a list of shares with exactly the right length to cause the problem.