i hate rewriting dce/rpc reassembly code.
this bug was due to not freeing up msrpc data after it had been
sent.  this wasn't a problem before i added the code that wanted
to send as many dce/rpc fragments as it could, rather than relying
on the smb client to "trigger" dce/rpc fragment reading / writing.