more clean-out to use local variables instead of ones in rpcsrv_struct.
this is in preparation to create a DCE/RPC security api.

also, have to use read_with_timeout, not read_data.  this _may_ work,
i'm really not sure, to be honest, but read_data() certainly doesn't
work, to read from the unix msrpc daemon socket.  you just can't
guarantee when an SMBtrans comes in how much data is going to come
back.  it's _at most_ the maximum SMBtrans receive size, that's all
you know.

returning _less_ than this is ok, because it's followed up by SMBreadX
calls, which request a specific amount.

so what i did was to request a _minimum_ of 16 bytes in read_with_timeout()
so that the client can interpret the DCE/RPC header and can issue
SMBreadX(s) for the rest of the fragment.