created a "no authentication" server-side instance of dce/rpc authentication. code in rpc_server/srv_pipe_srv.c is now reduced to handling dce/rpc packets, whilst decoding them is done in auth-instance modules.
the server-side authentication API has five functions, and one of those
is a bit of an odd-ball.  it obtains the user session key, which is
currently only ever used in srv_samr.c but WILL also be used in srv_lsa.c
for set / retrieve private data when i get round to implementing it.

i _may_ remove this odd-ball, and pass user_sess_key as an argument to
one of the other functions.