excellent!!!  my first use of tdb, and it worked!
in order to get round two problems:

1) secure netlogon channel opens TWO DCE/RPC NETLOGON connections, one
to negotiate the channel, the other to do secure netlogons over.

2) TCP connections can be DROPPED but the NETLOGON credential chain MUST
still be maintained.  prior implementations would _lose_ the credential
chain because smbd would terminate, and so would the msrpc netlogon
service, taking the local stack copy of the creds with it.

... i created netlogoncreds.tdb which stores struct dcinfo.  the key
is "DOMAIN_NAME\0SERVER_NAME".

there is a login DoS attack because of this:

it will be possible to just fake up a domain name and server name, and
send masses of fake NetrReqChallenge() followed by NetrAuthenticate2()
requests, which will *overwrite* the entry saved by valid workstations.

i'm... not really sure how to deal with this, yet.

...by the way, NT is likely to have the same DoS, if my suspicions are
correct [that they use a similar design, due to the nature of the
problems 1) and 2), above]