this one's a _great_ fun commit.
tracking down the user session key, getting it to work, deciding where
to _put_ it.  darn, i need to update cvs main, too, to make it compatible,
as i moved the usr_sess_key into CREDS_HYBRID.

also includes some clean-up / memory-related patches from Elrond, plus
a patch from him that uses auto-generated RIDs to identify the type,
so that users and groups are not confused when trying to decide if
an NT name is a user or a group (for LsaLookupNames responses).