i like this library thing so much i made a libsmbpw as well.  this contains passdb/smb*.c, passdb/sam*.c _and_ groupdb/*.c.
as a first implementation, it contains the stubs for the ldap, nisplus,
mysql _and_ the smbpasswd db code, but a _much_ better implementation
would be to define:

LIBLDAPPW

LIBMYSQLPW

LIBNISPLUSPW

LIBSMBPASSPW

and then select which one to link against.

hey, cool: that way, we can dlopen() libraries and do a conversion from
smbpasswd to, say, ldap, and back!

even MORE cool, we can do a backwards-compatible LDAP dlopen() and an
NT5 LDAP dlopen() and do an automatic conversion!