samr api conversion.  this one's a little bit more significant.
1) modified nterr.h to #define NT_STATUS_xxxx (0xC0000000 | errcode)

2) made a SAM_USERINFO_CTR and made it dynamically allocated.  elrond, help!

i realised that the user session key can be picked up from the POLICY_HND,
it doesn't need to come from the vuser struct!  i'm so stupid!

i've been worrying about this for ages.  i was thinking, how am i going to
make all the _samr_xxx() functions match up with the samr_xxx() functions
when i need to get state-info such as the user session key?  well, it's
simple.  instead of passing it as a function parameter to _samr_xxx(),
you store it as state-info associated with the POLICY_HND argument,
which references the cli_connection state info, which has the session key.

yesss :)

jean-francois, you may want to consider doing the same thing for the spoolss
code.  i will probably do this for you [or put some stubs in].