_spoolss_enumprinters spoolss conversion.
jean-francois, i think the BUFFER* argument and the PRINTER_INFO_CTR*
arguments are one and the same.  also, i duplicated some of the
buf_size, offered and returned parameters because i cannot identify
which ones are [in out], which are [in] and which are [out].

the problem is compounded by microsoft's poor design of the IDL definition
for this function, and it also can be used to send a REALLY large BUFFER*
of over 1mb of data, and cause an exception in spoolss.exe.

but that's another story...