made "enable" and "disable" the default "lpresume" and "lppause" commands for SYSV instead of the lpc commands (since these are BSD commands)
I still don't like the default "lppause" and "lpresume" commands for
SYSV since these seem to be SUN specific additions to the lp command
(at least I don't find them in IRIX and HPUX). I think this should be
inside an ifdef SUN??? but I'm not sure what the define should be so
I have left it for now.