from elrond.  this patch is NOT as large as it looks.
we're trying to subdivide the code down into independent libraries,
and that includes first the prototypes (starting with rpc_parse_proto.h)
and then moving on to the structures.

the idea is to be able to do --enable-maintainer-mode and not damn well
have EVERYTHING recompile if a make proto is done!