first stage of making some sort of "thread" context ... er, thing :) in the msrpc code.  it's not really threads, it's just a thread context, so that if different thread-contexts are requested, then the msrpc daemon will at least be able to switch user-security-context.
eventually, i will have to go so far as to be able to reconstruct PDUs
depending on the user context, but that will require one socket per
thread-context, and some means to set that up *inside* the Bind/Bind-Request
processing code [argh!], because that's what triggers a "new" user-context,
really.

i hope.