this weird-looking code uses the user_struct* that was obtained from an SMBsesssetupX to provide a response to api_WkstaUserLogon and api_NetUserGetInfo.
IF my suspicions are correct, an api_NetUserGetInfo or api_WkstaUserLogon
against an anonymous IPC$ connection will produce a failure, resulting
in the Win9x client DROPPING the anonymous connection and re-establishing
an authenticated SMBsesssetupX.

this will cause the smbd server to get a NET_USER_INFO_3 structure from
either the PDC or from itself (MSRPC remote or MSRPC loop-back, doesn't
matter which, it's all the same to domain_client_validate()), thence the
api_Net... or api_Wksta... call to follow will be provided with the correct
user logon info.

it also means that we can start filling in some of the "stub" fields,
such as last_logon_time, with _real_ info.  well, real, if netlogond
bothered to fill it in, but you get the idea :)