updated rpctorture.
as a result, found a client-connection (non-reuse) problem.  it's going
to require a minor redesign of the MSRPC client-connection architecture.
basically, i'm going to have to split "establish-connection-and-rpc-bind"
from "get-me-a-connection".  at present, "get-me-a-connection" also
does the rpc-bind bit, thereby producing potentially hundreds of msrpc
connections.

given that the samba msrpcd architecture forks() per connection, this isn't
such a good idea!