this one serious-attitude bug-fix.
ok.

rpctorture showed that msrpc connections not being reused wasn't good
enough.  results: one smbd process taking in several connections
to \PIPE\NETLOGON, and there's a deliberate hard-limit of 64 pipes
at the moment.

so....

i rewrote the msrpc connection code so that if there is a pre-existing
msrpc connection *with the same user-context (vuser_key)*, it is reused
that means NO bind-request because it will already have been done.

what took me so long was that the user-contexts (vuser_key) had to be
checked.  the architecture at the moment sends vuser_key on the dce/rpc
bind assoc_gid and dce/rpc header context_id fields.  unfortunately,
these were going over-the-wire on SMB connections, which they should NOT
do, it's supposed to only be used on ncalrpc (local rpc) unix-socket
loopback.

oops.

it's dog-slow... BUT IT WORKS!

i now have to find a way to call cli_net_use_add() and then actually
get the damn code to _use_ the client-side connection!

i think what is happening is that _every_ single new msrpc pipe initiated
gets its own SMB session established, which is not very good!