ok.  this introduces the following structures:
msrpc_local and ncalrpc_l_use.
ncacn_np and ncacn_np_use.

basically, what is happening is that connections can be reused (ncalrpc
for unix socket loop-back and ncacn_np for named pipes over SMB).

so, that's what these are for.

unfortunately, what makes this so stupidly complicated is that ncacn_np
can reuse SMB connections, not just SMB pipes.  so it uses the following
structures:

cli_state and cli_use.