ok.  a test previously showed that joining an NT5rtm to a domain with samrtdbd was failing, because the domain was opened read-only and then the user was opened read-write for create.
_un_fortunately, i had created an entire domain-user database, which of
course was opened read-only, and then the create-user OF course, failed.

so i had to change this to a system of creating INDIVIDUAL user databases
ARGH.

now, it should be possible to open at the domain-level (which is just a
directory), and then open a user for write-create, and away we go.