tim, after speaking with andrew he noticed that you were calling become_root() and unbecome_root().  client-side code CANNOT call these functions, it must only call the *stub* versions of these.
the only circumstances under which it is ok for client-side code to call
become_root() and unbecome_root() is if you are actually calling client-side
code from inside a root process to the msrpc loop-back interface.

this is because you cannot _access_ the msrpc loop-back interface unless you
are root (use a server name of "\\." instead of "\\someservername").