From Elrond@Wunder-Nett.org Tue Feb 29 05:53:10 2000 Date: Mon, 28 Feb 2000 17:29:48 +0100 From: Elrond <Elrond@Wunder-Nett.org> To: Luke Kenneth Casson Leighton <lkcl@samba.org> Subject: _samr_query_sec_obj
Hi Luke,

- moved the SD-generation code out of
  samrd/srv_samr_sam_tdb.c:_samr_query_sec_obj into its own
  samrd/samr_util.c
- updated the SD to have the same order and include
  Builtin\Account Operators, as seen on an nt4sp4 ws.
- used this function in samrpassd (which was broken here
  and usrmgr didn't work) and samrtdbd (for consistency)
  (I can't compile-check samrnt5ldap, so I will send Luke
  Howard a patch to inspect.)
- clean-up: Mainly policy-handles... but also other stuff
- debugging and type-checking of data associated with
  POLICY_HND in samrtdbd
  (I suspect, there's a problem in sam_query_sec_obj, see
  the new WARNING, that appears.)
- rpcclient/samedit: enumaliases has new option:
  [-d Domain], so I could do "enumaliases -d Builtin"


After this patch my usrmgr worked better, but now it looks
like it is thinking, it talks to a WS instead of a server.
I can't see the domain-groups, only aliases, and in the
user-properties, there are only 3 buttons for "Groups",
"Profiles" and "RAS", the others are missing... I tried a
long time to find out about it... I failed, so could you
have a look at this?

[lkcl: this because winregd is returning "i am a wksta"]

Okay, and another thing:

You seemed to look for the bit on the user-objects, that
allows password-changing: I think, it must be 0x0004,
cause:
Everyone has: 0x2035b
Normal user : 0x20044

0x20040 is included in the bits, everyone has, 0x0004 is
left over....

Just some thoughts...


    Elrond