From Elrond@Wunder-Nett.org Fri Mar 17 08:34:21 2000 Date: Thu, 16 Mar 2000 16:29:53 +0100 From: Elrond <Elrond@Wunder-Nett.org> To: Luke Kenneth Casson Leighton <lkcl@samba.org> Subject: SVC_UNKNOWN_3
Hi Luke,

I played a bit with srvmgr around and it did a opnum=3 on
svcctl, so I tried to decode the request. For me it looked
lit it only consists of a sc_man_hnd. So I added some
parsing code and an "svcunk3" to rpcclient, so I could ask
the NT-box for the answer. The answer were 20 NUL-bytes. I
guess, this is a closed handle and a status. Also the
following svc_close(sc_man_hnd) seemed to fail somehow (I
didn't realy see the fault-pdu or so... but I did not see
an svc_io_r_close_hnd either.)

[lkcl: i suspect that this is a "delete service" call,
       which is why the close fails because the handle
       is alread closed!]

All this looks to me like opnum=3 is something like
SVC_CLOSE2 or so.

Could you apply the patch and verify that?


I also did the following:
- modified configure.in, so it is more like HEAD (still lot
  to do)
- fixed some issues in groupdb/builtinunix.c:
  When I created aliases with domain alias map, all these
  aliases also existed in the Builtin domain, so usrmgr
  showed them all doubled. (might be interesting for
  HEAD...)
- made global_sid_S_1_1_0 static, one should use
  global_sid_everyone instead.
- continued my share_get_info-effort by writing *_free_*
  for all the fun. Yet have to use them.


    Elrond