From Elrond@Wunder-Nett.org Sat Mar 18 04:49:11 2000 Date: Fri, 17 Mar 2000 16:02:28 +0100 From: Elrond <Elrond@Wunder-Nett.org> To: Luke Kenneth Casson Leighton <lkcl@samba.org> Subject: Re: SVC_UNKNOWN_3
On Fri, Mar 17, 2000 at 08:27:49AM +1100, Luke Kenneth Casson Leighton wrote:
> it's probably a delete service :-) :-)
>
> check your box!!!
[...]

Hi Luke,

You pass in a Service-Control-Manager-Handle, not a
service-handle, so it can't be service-delete. It could of
course be scman-del... which would be quite crazy.
(includind the fact, that this is done by srvmgr, which
hasn't any means to add/del a service...)

But it was all my fault. I copied cli_svc.c:svc_close()
over to svc_unknown_3... and guess what... I changed
everything except the Opcode...

After that I was more successful: You pass in a scman_hnd,
and get back some other handle...I don't know yet, what it
is good for... I added an appropiate server-side
implementation and srvmgr.exe got a bit further, but now it
calls opcode 11 on a service-handle. Maybe more soon.
(I'm pressing the "Startup"-button in Services in srvmgr)


The patch contains:
- complete server and client-side implementation of
  SVC_UNKNOWN_3 (yet no idea, what it is, but we have it)
- --with-privatedir=... from HEAD is now completely in TNG
- merged some things from HEAD into lib/time.c
  PLEASE DON'T run indent on lib/time.c !
[lkcl: okie]
- dito for param/loadparm.c
  PLEASE DON'T run indent on param/loadparm.c !
[lkcl: okie]


    Elrond