ok, this one's weird.  firstly, sam_set_userinfo doesn't align the 2nd info level at the container, but sam_set_userinfo2 *does*.
secondly, it _looks_ like there are two revisions for SAM database
entries, where some unknown trash is _not_ placed in the structure
(at the end: unknown_6 and padding4) if the revision number (unknown_5)
is zero.

have to keep an eye on this one over the next year or so, to make sure
all cases are caught.

third, it _looks_ like if you do a password set that the account disabled
acb_info bit is automatically unset, but i cannot confirm that right
now, i'll just have to wing-it (in the next commit)