disabled the getgrent() code recently added that returns all unix groups as NT aliases.  no allowance is made for possible clashes with some unix groups that may be mapped to NT groups, not NT aliases.
it is not the place of this code to make decisions about which unix
groups should be mapped to NT aliases or NT groups, or in fact whether
unix groups exist at all (e.g "appliance" mode).

please could the person who wrote this code contact me and i will outline
the architecture and design issues that you need to be aware of when
working in this area: i need your skills and ability -- applied in the
right places.

thank you.