JF - you might want to look at this patch.
Went through and fixed up all failure possibilities. If something (malloc,
file open etc.) fails, then this code goes through and cleans up all allocated
resources on exit. This is the equivalent of a C++ destuctor for the allocated
structures. We need code to be written like this to pass purify tests.

Jeremy.