uhh, guys: i spent my entire day updating from yesterday's cvs main commits.  i mean, nothing else. i wasn't able to do any debugging of the issues from yesterday.
so now, there are two possible sets of bugs, compounded.  this
commit compiles but hasn't been tested, i haven't had a chance
to.

i'd really appreciate it if this stopped happening.  once is enough.

now i am going to get something to eat and if i am lucky, i will be
able to hold a fork.  this is serious.

now i've updated TNG to cvs main, don't make me maintain it, please
do it yourselves.

thanks.

p.s i'm never, ever, working in a branch again.  if that's not ok,
i won't do the work.  it's not worth it.